Usnadni použití capitalize nejen jako filtru v šabloně
